requirements.txt
google-generativeai
openpyxl 
pandas 
gspread 
oauth2client
streamlit
python-dotenv